AI-Powered-Product-Finder/Product-Finder-Alloydb/Backend/Product-Finder/requirements.txt
fastapi==0.121.2
google-cloud-alloydb-connector==1.9.1
pandas==2.3.3
psycopg2-binary==2.9.10
pg8000==1.31.5
pydantic==2.9.2
python-dotenv==1.1.1
SQLAlchemy==2.0.44
uvicorn==0.35.0